Get the marginal cumulative incidence in the treated and untreated groups based on a Cox model or Kaplan Meier curve.

#' @description Internally calls compute_cox_ve() or compute_km_ve() depending on the value of method

Usage

compute_marginal_ve(
  adata,
  adata_outcome_name,
  adata_event_name,
  adata_trt_name,
  eval_times,
  method,
  adjust = NULL,
  censor_time = max(eval_times),
  separate = TRUE,
  keep_models = TRUE
)

Arguments

adata

A data frame that represents the analysis data set of a clinical trial.

adata_outcome_name

Character string specifying the time to event variable in adata. The time should be the time to event from vaccination/matched index date

adata_event_name

Character string specifying the event variable in adata

adata_trt_name

Character string specifying the treatment variable in adata

eval_times

Numeric vector specifying the timepoints at which to compute cumulative incidence and the derived effect measures. The timepoints should be expressed in terms of time since exposure. All values must be greater than tau and and should correspond to clinically meaningful follow-up durations, such as 30, 60, or 90 days after exposure. A fine grid of timepoints (e.g., eval_times = (tau+1):100) can be provided if cumulative incidence curves over time are desired.

keep_models

Logical; return the two fitted hazard models used to compute cumulative incidences? Default: TRUE.

Value

A list containing the following:

estimates

A matrix of estimates. The columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.

models

If keep_models = TRUE, the models used to compute risk and VE