Helper to estimate generic bootstrapped confidence interval

This function computes Wald and percentile bootstrapped confidence intervals for the proposed VE estimator.

Usage

estimate_bootstrap_ci(
  one_boot_function_name,
  one_boot_args,
  ci_type,
  boot_reps,
  pt_est = NULL,
  alpha = 0.05,
  keep_boot_samples = TRUE,
  n_cores = 1
)

Arguments

one_boot_function_name

Character string naming the function that computes one bootstrap iteration of estimates

one_boot_args

List of arguments to pass to the function that computes one bootstrap iteration of estimations

ci_type

Method for constructing bootstrap confidence intervals. One of "wald", "percentile", or "both".

"wald" (default): Computes Wald-style intervals using bootstrap standard errors. See Confidence intervals section for details.

"percentile": Computes percentile bootstrap intervals.

"both": Computes and returns both sets of intervals.

boot_reps

Number of bootstrap replicates for confidence intervals. Recommended to use at least 1000 for publication-quality results. Use smaller values (e.g., 10-100) for initial exploration. Default: 0 (no bootstrapping).

pt_est

A matrix of estimates. The columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.

alpha

Significance level for confidence intervals (Confidence level = 100*(1-alpha)%). Default: 0.05.

keep_boot_samples

Logical; return bootstrap samples? Default: TRUE. Must be set to TRUE if user plans to use add_simultaneous_ci() to obtain simultaneous confidence intervals.

n_cores

Integer; parallel cores for bootstrapping. Passed to parallel::mclapply as mc.cores. On Unix-like OS only; not available on Windows. Default: 1.

Value

A list containing the following:

ci_estimates

A list of matrices containing the lower and upper confidence interval bounds and bootstrap standard error for each term. When ci_type = "wald", the bootstrapped standard errors on the transformed scale are also included.

n_success_boot

A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are those that result in non-missing valid point estimates.)

boot_samples

If keep_boot_samples = TRUE, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and the columns are the time points.