Fit survival models to estimate hazards and predict risk

fit_model_0() fits a Cox model to estimate risk for unvaccinated individuals

this model includes all individuals, censoring vaccinated individuals at their time of vaccination

TODO: need to censor individuals at d_max + t0? d_max not that well defined.

fit_model_1() fits a Cox model to estimate risk for vaccinated individuals

this model only includes vaccinated individuals at risk tau days after their vaccination

furthermore, individuals are censored at t0 days after vaccination

Usage

fit_model_0(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  formula_0 = NULL
)

fit_model_1(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  censor_time,
  tau,
  formula_1 = NULL
)

Arguments

data

A data frame containing pertinent information for VE estimation.

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

adjust_vars

A character vector containing the names of variables in data to adjust for.

formula_0

A formula or pre-fit model for estimating hazards in unvaccinated group (recommended for use only by power users)

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

formula_1

A formula or pre-fit model for estimating hazards in vaccinated group. (recommended for use only by power users)

Value

The fitted survival object