Get empirical probability distributions g(d|x) and p(x) based on input data

Get empirical probability distributions g(d|x) and p(x) based on input data

Usage

get_gp(df, outcome_time, exposure, exposure_time, covariates, tau)

Arguments

df

A data frame representing the target population for the estimated distributions

typically, population involves the "analysis-eligible" subset of a given data source

outcome_time

Name of the time-to-event/censoring variable. Time should be measured from a given time origin (e.g. study start, enrollment, or age) for all individuals.

exposure

Name of the exposure indicator. The underlying column should be numeric (1 = exposed during follow-up, 0 = never exposed during follow-up).

exposure_time

Name of the time to exposure, measured from the chosen time origin; use NA if not exposed. Time must be measured in the same units (e.g. days) as that used for outcome_time.

covariates

Character vector of covariates to adjust for when fitting the hazard models. These covariates should include all known confounders of exposure and censoring measured at the chosen time origin.

tau

Numeric. Only vaccinated who are at-risk tau days after vaccination are included.

Value

A list of the following:

g_weights

A data frame containing the covariate-conditional day probabilities. Columns include group_id identifying the covariate group, <exposure_time> specifying the day, probspecifying the covariate-conditional day probability, and each variables incovariates} \item{p_weights}{A data frame containing the covariate probabilities. Columns include group_id identifying the covariate group,probspecifying the covariate probability, and each variable incovariates`