Compute VE point estimate in matched data set

First, creates the analysis matched data set based on matched_data. Then calls compute_marginal_ve() on the analysis data set, specifying the appropriate outcome and treatment arguments.

Usage

get_one_matching_ve(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = TRUE,
  return_models = TRUE
)

Arguments

matched_data

A data frame for the matched cohort

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

method

Character string specifying method for survival estimation ("cox" for Cox proportional hazards regression model or "km" for Kaplan-Meier estimation)

adjust

If method = "cox", a formula or character vector containing the names of covariates in matched_data to adjust for.

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

pair_censoring

Logical to indicate if matched pairs should be censored when unvaccinated pair becomes vaccinated.

separate

If method = cox and separate = TRUE, Cox regression models are fit in the treated and untreated groups separately.

return_models

Logical value: Should fitted survival models be returned?

Value

The object returned by compute_marginal_ve()