Compute VE point estimate

This function computes point estimates of vaccine effectiveness based on the proposed method. Internally called by nomatchVE and one_boot_ve.

Usage

get_one_ve(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  matched_dist_options,
  times,
  censor_time,
  tau,
  formula_0,
  formula_1,
  return_models = TRUE,
  return_gp_list = TRUE,
  return_matching = TRUE
)

Arguments

data

A data frame containing pertinent information for VE estimation.

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

adjust_vars

A character vector containing the names of variables in data to adjust for.

marginalizing_dist

Generally, should be set to "observed", the default value, which estimates the marginalizing distributions based on the observed data. For historical purposes, marginalizing_dist can also be set to "matched", which uses the marginalizing distributions of a matched dataset and requires use of the matched_dist_options argument. To providing a fixed choice of marginalizing weight functions, marginalizing_dist should be a list of 2 dataframes named g_dist and p_dist. g_dist must be a dataframe with the columns <time_name> describing the time of vaccination, all variables in , group_name identifying each unique covariate group, and prob for the corresponding probability. p_dist must be a dataframe with all variables in and variable prob for the corresponding probability.

matched_dist_options

If marginalizing_dist == "matched", a list of parameters controlling process of creating matched cohort and matched analysis dataset from which marginalizing distributions will be estimated. List can be created by matched_dist().

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

formula_0

A formula or pre-fit model for estimating hazards in unvaccinated group (recommended for use only by power users)

formula_1

A formula or pre-fit model for estimating hazards in vaccinated group. (recommended for use only by power users)

return_models

Logical: Should survival models be returned?

return_gp_list

Logical: Should marginalizing distributions be returned?

return_matching

Logical: Should matched datasets be returned? Default is TRUE if marginalizing_dist = "matched" and FALSE otherwise.

Value

A list containing the following:

estimates

A matrix of estimates where the the columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.

model_0, model_1

If return_models = TRUE, the fitted survival models for unvaccinated and vaccinated.

gp_list

If return_gp_list = TRUE, the list of marginalizing distributions used.

matched_data

If return_matching = TRUE, the matched cohort

matched_adata

If return_matching = TRUE, the analysis-eligible matched cohort