Perform matching and get matched adata back

Perform matching and get matched adata back

Usage

match_adata(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  id_name,
  matching_vars,
  replace = FALSE,
  seed = NULL,
  tau,
  pair_censoring = TRUE
)

Arguments

data

A data frame for which matches are sought

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string representing the original event variable in matched_data

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

id_name

Character string specifying name of a unique individual identifier in data.

matching_vars

A character vector containing the names of variables in data to match on .

replace

Logical: Should matching be done with replacement?

seed

Numeric: seed to set prior to performing random matching

tau

The time excluded after vaccination to allow building up of immunity

pair_censoring

Logical to indicate if matched pairs should be censored when unvaccinated pair becomes vaccinated.

Value

A list containing the matched analysis dataset and the matched cohort